requirements/develop.txt
Django==1.6.7
GraphLab-Create==0.3.0
Parsley==1.1
Pillow==2.4.0
PyYAML==3.10
Shapely==1.5.13
South==1.0.2
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.17
boto==2.29.1
celery==3.1.11
coverage==3.7
coveralls==0.3
decorator==3.4.0
django-appconf==0.6
django-celery==3.1.10
django-compressor==1.4
django-guardian==1.2.0
django-leaflet==0.17.0
django-maintenancemode==0.10
django-picklefield==0.2.1
django-s3-folder-storage==0.1
django-ses==0.4.1
django-storages==1.1.5
django-su==0.4.2
django-tinymce==1.5.2
django-userena==1.3.0
django-zebra==0.4.4
djangorestframework==3.1.3
docopt==0.6.1
easy-thumbnails==1.5
geojson==1.3.1
html2text==2014.4.5
httplib2==0.7.7
jsonfield==0.9.20
kombu==3.0.16
librato-metrics==0.4.9
lucene-querybuilder==0.2
mixpanel-py==3.1.1
neo4jrestclient==2.0.3
networkx==1.8.1
numpy==1.8.1
pandas==0.14.0
psutil==1.1.3
psycopg2==2.4.5
pyblueprints==0.5.1
pycrypto==2.6
pydkim==0.3
python-dateutil==2.2
python-rexster==0.1.1
pytz==2014.3
pyzmq==14.3.0
requests==2.3.0
selenium==2.42.1
sh==1.09
simplejson==2.6.2
six==1.6.1
splinter==0.6.0
stripe==1.7.8
ujson==1.33
unicodecsv==0.11.0
wsgiref==0.1.2
